[#15] Split main app into container + presentation components.
